+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 7.0.0-beta.23
+
+- The internal `useMain` hook has been renamed to `useInstance`
+- The internal `useBeforeDimensions` hook has been renamed to `useInstanceBeforeDimensions`
+
 ## 7.0.0-beta.22
 
 - Fixed an issue where `useRowState` would crash due to invalid initial state attempting to spread into the new state